Drops ViewHelper, adds resource entry for clear animation duration, minor refactoring/reformatting
