+++ scratchoff/src/main/res/values/integers.xml
@@ -1,5 +1,6 @@
 <resources>
 
     <integer name="scratch__threshold_percent">65</integer>
+    <integer name="scratch__clear_animation_duration_ms">1000</integer>
 
 </resources>
